LayoutTests/touchadjustment/event-triggered-widgets.html

<html>
<head>
<style>
    .box { border: 1px solid black; border-radius: 5px 5px; margin: 1em; width: 40em; }
</style>

<script src="../fast/js/resources/js-test-pre.js"></script>

</head>

<body onload="runTests()">

<p id='description'></p>

<div id='console'></div>

<script>
    var element;
    var adjustedNode;
    function findAbsolutePosition(node) {
        var pos = new Object();
        pos.left = 0; pos.top = 0;
        do {
            pos.left += node.offsetLeft;
            pos.top += node.offsetTop;
        } while (node = node.offsetParent);
        return pos;
    }

    function findBestClickableNode(x, y, width, height)
    {
        var adjustedNode = internals.touchNodeAdjustedToBestClickableNode(x, y, width, height, document);
        if (adjustedNode && adjustedNode.nodeType == 3) // TEXT node
            adjustedNode = adjustedNode.parentNode;
        return adjustedNode;
    }

    function testDirectTouch(element)
    {
        var pos = findAbsolutePosition(element);
        var x = pos.left + element.clientWidth / 2 - 1;
        var y = pos.top + element.clientHeight / 2 - 1;
        var width = 3;
        var height = 3;
        adjustedNode = findBestClickableNode(x, y, width, height);
    }

    function testDirectFatFinger(element)
    {
        var pos = findAbsolutePosition(element);
        var x = pos.left + element.clientWidth / 2 - 1;
        var y = pos.top - 5 ;
        var width = element.clientHeight;
        var height = element.clientHeight + 10;
        adjustedNode = findBestClickableNode(x, y, width, height);
    }

    function testIndirectFatFinger(element)
    {
        var pos = findAbsolutePosition(element);
        var x = pos.left + element.clientWidth / 2 - 1;
        var y = pos.top - 7;
        var width = 10;
        var height = 10;
        adjustedNode = findBestClickableNode(x, y, width, height);
    }

    function testTouchHit(elementid, touchType) {
        element = document.getElementById(elementid);
        touchType(element);
        shouldBe('adjustedNode.id', 'element.id');
    }

    function testTouchMiss(elementid, touchType) {
        element = document.getElementById(elementid);
        touchType(element);
        shouldNotBe('adjustedNode.id', 'element.id');
    }

    function testDirectTouches()
    {
        debug('Testing small direct hits');
        testTouchHit('test1', testDirectTouch);
        testTouchHit('test2', testDirectTouch);
        testTouchHit('test3', testDirectTouch);
        testTouchMiss('test4', testDirectTouch);
        testTouchHit('test5', testDirectTouch);
    }

    function testDirectFatFingers()
    {
        debug('Testing large direct hits');
        testTouchHit('test1', testDirectFatFinger);
        testTouchHit('test2', testDirectFatFinger);
        testTouchHit('test3', testDirectFatFinger);
        testTouchMiss('test4', testDirectFatFinger);
        testTouchHit('test5', testDirectFatFinger);
    }

    function testIndirectFatFingers()
    {
        debug('Testing large direct hits');
        testTouchHit('test1', testIndirectFatFinger);
        testTouchHit('test2', testIndirectFatFinger);
        testTouchHit('test3', testIndirectFatFinger);
        testTouchMiss('test4', testIndirectFatFinger);
        testTouchHit('test5', testIndirectFatFinger);
    }

    function runTests()
    {
        if (window.layoutTestController && window.internals && internals.touchNodeAdjustedToBestClickableNode) {
            description('Test various ways to trigger input-widgets. On a touch interface, all the actions should be triggerable with either a touch down or a touch tap.');
            layoutTestController.dumpAsText();
            layoutTestController.waitUntilDone();
            testDirectTouches();
            testDirectFatFingers();
            testIndirectFatFingers();
            isSuccessfullyParsed();
            layoutTestController.notifyDone();
        }
    }
</script>

<script>
    function triggerInput() {
        var element = event.srcElement;
        if (!element.open) {
            element.innerHTML = '<input type=text style="width: 100%"></input>'
            element.open = true;
        }
        element.firstChild.focus();
    }
</script>

<div id=test1 class=box tabindex=1 onfocus='triggerInput()'>
Focus here should give a text input-field.
</div>

<div id=test2 class=box onmouseover='triggerInput()'> 
Mouse-over here should give a text input-field.
</div>

<style>
    .box:not(:hover) .hovertriggered { visibility: hidden;}
    .box:hover .hoverfallback { display: none; }  
</style>

<div id=test3 class=box> 
    <span class=hoverfallback>Hovering here should give a text input-field.</span>
    <input type=text class=hovertriggered></input>
</div>

<div id=test4 class=box onfocus='triggerInput()'>
    <span tabindex=1> Focusing here should only give focus outline to this inner text.
    </span>
</div>

<div id=test5 class=box>
    <span tabindex=1> Focusing here should give a text input-field.
    </span>
</div>

<script>
    var element = document.getElementById('test5');
    element.addEventListener('DOMFocusIn', triggerInput, false);
</script>

</body>
</html>
