LayoutTests/touchadjustment/event-triggered-widgets-expected.txt

Test various ways to trigger input-widgets. On a touch interface, all the actions should be triggerable with either a touch down or a touch tap.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Testing small direct hits
PASS adjustedNode.id is element.id
PASS adjustedNode.id is element.id
PASS adjustedNode.id is element.id
PASS adjustedNode.id is not element.id
PASS adjustedNode.id is element.id
Testing large direct hits
PASS adjustedNode.id is element.id
PASS adjustedNode.id is element.id
PASS adjustedNode.id is element.id
PASS adjustedNode.id is not element.id
PASS adjustedNode.id is element.id
Testing large direct hits
PASS adjustedNode.id is element.id
PASS adjustedNode.id is element.id
PASS adjustedNode.id is element.id
PASS adjustedNode.id is not element.id
PASS adjustedNode.id is element.id
PASS successfullyParsed is true

TEST COMPLETE
Focus here should give a text input-field.
Mouse-over here should give a text input-field.
Hovering here should give a text input-field. 
Focusing here should only give focus outline to this inner text.
Focusing here should give a text input-field.
